Allow getClassName to take an options argument
The only option is to silence styleName errors, for example in production.
If `silenceStyleNameErrors` is true, the styleNames which cannot be
found will simply be removed from the resulting classNames, matching
the behavior of `react-css-modules`.

This commit alone will not result in a change in behavior, the next
steps will be to pass in the option from the babel-plugin.