feat: add `handleMissingStyleName` option (#112)
* Add .eslintignore for test fixtures

The double quotes that the plugin adds go against the linting 
config of the project, and editors that auto-fix linting rules
will try to convert quotes in `expected.js` files to single quotes,
causing tests to fail.  It seems safest to ignore all fixture files
from linting.

* Allow createObjectExpression to take a boolean

* Allow getClassName to take an options argument

The only option is to silence styleName errors, for example in production.
If `silenceStyleNameErrors` is true, the styleNames which cannot be
found will simply be removed from the resulting classNames, matching
the behavior of `react-css-modules`.

This commit alone will not result in a change in behavior, the next
steps will be to pass in the option from the babel-plugin.

* Add `silenceStyleNameErrors` to options schema

* Pass silenceStyleNameErrors to getClassName

This hooks up the options set by the user into getClassName, so that it
has an effect.  However, getClassName can also be used at runtime,
which this commit does not address.

* Provide silenceStyleNameErrors at runtime

This commit adds the options object for getClassName iif silenceStyleNameErrors
is true.  By not always including the options object, it helps to save 
a few characters in the generated code.

* Restructure Configuration docs

Seeing the flow types first thing in the Configuration section was
confusing to new users, so this adds a brief example of how to set
babel plugin options, and moves the flow types underneath the table
of options.

* Add silenceStyleNameErrors to README

* Loosen InputObjectType

Flow was already not happy with using `InputObjectType` within its own
definition, and was converting it to `{[key: string]: string | any}`
when I inspected it in Nuclide.  

I’m honestly not sure how to specify a union type on an indexer as was
being attempted, because flow complains as soon as you try to provide
an object that has a property with a typed value.

* Allow more options to handle missing styleNames

This changes the option from a boolean “silenceStyleNameErrors” to an
enum “handleMissingStyleName” option, with possible values of “throw”
(default), “warn”, and “ignore”.